Split terminal integration tests from workspace model
